backend/voice/requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
python-socketio==5.11.0
Flask-Cors
google-cloud-speech>=2.21.0
google-cloud-texttospeech>=2.14.0
pydub>=0.25.1
numpy>=1.24.0
gunicorn
eventlet  # For production WebSocket